requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0,<4.0
djangorestframework-simplejwt>=5.2.2,<6.0
PyJWT>=2.6.0,<3.0
pytz
sqlparse
django-cors-headers>=3.13.0
